products data fetch with api and use hooks useState useEffect
