+++ frontend/src/screens/HomeScreen.jsx
@@ -1,8 +1,20 @@
+import { useEffect, useState } from "react";
+import axios from "axios";
 import { Row, Col } from "react-bootstrap";
-import products from "../products";
 import Product from "../components/Product";
 
 const HomeScreen = () => {
+    const [products, setProducts] = useState([]);
+
+    useEffect(() => {
+        const fetchProducts = async () => {
+            const { data } = await axios.get("/api/products");
+            setProducts(data);
+        };
+
+        fetchProducts();
+    }, []);
+    
     return (
         <>
             <h1>
