+++ frontend/src/screens/ProductScreen.jsx
@@ -1,12 +1,23 @@
+import { useState, useEffect } from "react";
 import { useParams } from "react-router";
 import { Link } from "react-router-dom";
 import { Row, Col, Image, ListGroup, Card, Button } from "react-bootstrap";
-import products from "../products";
 import Rating from "../components/Rating";
+import axios from "axios";
 
 const ProductScreen = () => {
+    const [product, setProduct] = useState({});
     const { id: productId } = useParams();
-    const product = products.find((p) => p._id === productId);
+
+    useEffect(() => {
+        const fetchProduct = async()=>{
+            const {data} = await axios.get(`/api/products/${productId}`);
+            setProduct(data);
+        }
+
+        fetchProduct();
+
+    }, [productId]);
     return (
         <>
             <Link className="btn btn-light my-3" to="/">Go Back</Link>
